Create and extend abstract LoaderTest.
